us-stock-assistant/mcp-servers/market-data-server/requirements.txt
mcp[cli]>=1.26.0
httpx>=0.28.0
python-dotenv==1.0.0
